==> bookdown/05-Seasonal_validation.rmd <==
# Validation

```{r setup, echo = F, message = F}
knitr::opts_chunk$set(fig.width = 5)
knitr::opts_chunk$set(fig.height = 6)
library(SeaVal)
library(ggpubr)
library(ggplot2)
load(file = '_temp.RData')
load(file = '_data_dir.RData')

```


In this section we give examples how to evaluate seasonal forecasts with SeaVal. One of the key tasks addressed by `SeaVal` is the evaluation of tercile probability forecasts, but also other forecast types are supported. We recommend to read the [WMO recommendations on evaluating Seasonal Climate Forecasts](https://library.wmo.int/doc_num.php?explnum_id=4886), which discusses in detail aspects of evaluating tercile forecasts.


## Evaluating Tercile Forecasts

One of the main products disseminated at GHACOFs are probabilistic forecasts whether the coming season will see a below normal-, normal-, or above normal amount of rainfall. 
These three categories are defined by climatological terciles, below normal meaning that the the year was dryer than 2/3 of all observed years in a climatological reference period.
Therefore, we call these type of predictions tercile forecasts. 
Typically, high emphasis is put on the 'most likely category', according to the prediction. For example, the typical probability maps show for each gridpoint only the probability of this category.

From an evaluation perspective, however, it is preferable to not only consider the most likely category, but to evaluate the vector of three probabilities for all categories.
This vector conveys more detailed information about the forecast: Say, for example, two competing models predicted the probabilities (0.5, 0.3, 0.2) and (0.5, 0.45, 0.05), respectively (in the order below, normal, high). Say now, after observing the predicted season, it turns out that the rainfall was in fact above normal. In this case, both predictions were pretty bad, but the first model at least assigned a 20% chance to above-normal-rainfall, whereas the second model only assigned a 5% chance to that outcome. So the first prediction was substantially better than the second. However, if evaluation focuses on the category with the highest predicted probability, the two models can't be distinguished: For both of them this category is 'below' with 50\% probability.

Therefore, considering all three probabilities of the prediction allows for better forecast evaluation. This does not mean, however, that the communication of the prediction to the public needs to contain all three probabilities, which would likely be more confusing than helpful. 
In the next subsection we'll discuss how to evaluate a fully probability forecast (vector of three probabilities). 
In some scenarios, however, only the highest probability category (and its assigned probability) are communicated and available for evaluation. In Section @ref{eval-terciles2} we briefly discuss how forecasts can be evaluated in this case.

Throughout this section we will work with the data example derived [here](#us-obs).

### Verification maps ###

For getting a good idea whether a tercile forecast for a past season was good, it is a good idea to just plot the observation. And since we predict in the three categories below, normal and above, it makes sense to transform the observation to this scale and plot, where precipitation was unusual and how so. To this end we need to load more than one year of observation, in order to establish what constitutes normal rainfall for a location. This can be done by the function `ver_map`:

```{r }
# get observations (example chirps dataset for December):
dt_Dec = chirps_monthly[month == 12]

ver_map(dt_Dec,o = 'prec')

```

Here, the brown grid points show where precipitation was below normal, blue shows grid points with normal rainfall and green shows above normal rainfall, with darker brown/green indicating more extreme observations. The `ver_map` function by default considers all years available in the data table to calculate the reference climatology, and plots the verification map for the last available year. So the plot above is the verification map for December 2020. These defaults can of course be changed, see functiion documentation.

## Scores for full tercile forecasts {#eval-terciles}

By scores we mean any validation tool that returns a number. Typical examples are the ignorance score (or log-likelihood) or the ROC-score (or area under curve/AUC). Some scores such as the ignorance score directly indicate 'goodness' of the forecast in the sense that better forecasts achieve on average better scores. Such scores are called *proper*. Other scores, such as the ROC-score evaluate properties such as the resolution of the forecast, but are agnostic towards bias of the forecast. See the [WMO recommendations](https://library.wmo.int/doc_num.php?explnum_id=4886) for an overview of scores, their properties and how to use them.


### Proper scores and proper skill scores 

The ignorance score evaluates forecast performance, with lower values indicating better skill:

```{r }
dt = dt_tercile_forecast
print(dt)

# get the ignorance score
igs = IGS(dt,
          o = 'tercile_cat')
ggplot_dt(igs,high = 'darkgreen',low = 'purple',discrete_cs = TRUE, mn = 'IGS for MAM tercile forecast 2021')
```

A common difficulty when considering proper scores is that the actual value of the score is difficult to interpret. For this purpose, scores are often transformed into *skill scores*. Skill scores indicate whether a forecast performed better or worse than a climatological forecast. A skill score of above zero indicates that the forecast has more skill than climatology. A skill score of 1 indicates a perfect forecast. For deriving the ignorance skill score we can simply call the `IGSS` function

```{r }
# get the ignorance score
igss = IGSS(dt,
            o = 'tercile_cat')
ggplot_dt(igss,high = 'darkgreen',low = 'purple',discrete_cs = TRUE,midpoint = 0, binwidth = 0.2, mn = 'IGS for MAM tercile forecast 2021')
```

Green areas in the plot indicate locations where the tercile forecast performed better than climatology, purple worse. 

A second score similar to the ignorance score is the *Multicategory Brier Score (MB)*. Its main advantage over the ignorance score is that it is less sensitive to outliers and always is finite (whereas the ignorance score is $\infty$ when a zero-probability event occurs).
The MBS is defined as
\[\text{MB} :=  (p_1 - e_1)^2 + (p_2 - e_2)^2 + (p_3 - e_3)^2.\]
Here, $p_1,p_2,$ and $p_3$ are the predicted probabilities for the three categories, and $e_i$ is 1 if the observation falls in the $i$th category, and 0 else. For example, if the observation falls into the first category, the MB would be
\[(p_1 - 1)^2 + p_2^2 + p_3^2.\]

This score is strictly proper, meaning that it rewards calibration and accuracy. In our particular situation, the climatological forecast is uniform
(since climatology is used to define the tercile categories), and the climatological forecast (1/3,1/3,1/3) always gets a MB of 2/3. 
It is therefore very convenient to consider the *Multicategory Brier Skill Score (MBS)*
\[MBS := \frac{3}{2}(2/3 - \text{MB}).\]
Like other skill scores, this score is normalized in the sense that a perfect forecaster attains a skill score of 1 and a climatology forecast always gets a skill score of 0.
Note that, for the MBS, higher values indicate better performance, unlike for the MB (similar as for other scores such as MSE).
Since the MBS is just a linear transformation of the MB, but is way easier to interpret, there is not really a reason to ever calculate the MB, and `SeaVal` does not contain a separate function for that.
For calculating the  MBS, we can call the function with the same name:

```{r }
# get Multicategory Brier Score:
mbs = MBS(dt,
          o = 'tercile_cat')
ggplot_dt(mbs,high = 'darkgreen',low = 'purple',discrete_cs = TRUE,binwidth = 0.2,midpoint = 0, mn = 'MBS for MAM tercile forecast 2021')
```

As we can see, there is a large amount of agreement between the IGSS and the MBS.








To see whether the forecast was overall better than climatology, we average the MBS:

```{r}
# check out the MBS by country:
mbs = add_country_names(mbs)
mean_mbs = mbs[,.(mean_mbs = mean(MBS,na.rm = T)), by = country]
print(mean_mbs)

```

Finally, let's check whether this makes sense, by comparing climatology to the prediction:

```{r}
dt[,anomaly:= precip - clim]

ggplot_dt(dt[year == 2021],'anomaly',high = 'blue',low = 'red',midpoint = 0, mn = 'observed 2021 MAM precip anomaly')
```
```{r,fig.width = 20}
# or, as discrete plot:
pp1 = ggplot_dt(dt[year == 2021],'anomaly',
                high = 'blue',low = 'red',midpoint = 0,
                rr = c(-100,100),discrete_cs = TRUE,breaks = seq(-100,100,40),
                mn = 'observed 2021 MAM precip anomaly')

# also, let's plot the predicted probabilities:
pp2 = ggplot_dt(dt,'below',midpoint = 0.33,discrete_cs = TRUE,binwidth = 0.05,mn = 'predicted probability below')
pp3 = ggplot_dt(dt,'normal',midpoint = 0.33,discrete_cs = TRUE,binwidth = 0.05,mn = 'predicted probability normal')
pp4 = ggplot_dt(dt,'above',midpoint = 0.33,discrete_cs = TRUE,binwidth = 0.05,mn = 'predicted probability above')

ggpubr::ggarrange(pp1,pp2,pp3,pp4,ncol = 4)

```
As we can see, the season was very wet overall. The prediction was overall wet as well, especially over the western part of the considered region, where the prediction also got assigned a positive MBS.

### Evaluation when only the highest probability category is avaliable {#eval-terciles2}

As argued above, it is preferrable to evaluate tercile forecasts that are given as full probability vector containing all three probabilities. However, we might still face scenarios where we only have the highest probability category available, e.g. some older forecasts for which only this has been saved. What can we do in this case?

Intuitively, a promising candidate for a proper score seems to be the two-category-Brier score on the category with the highest probability
\[BS_{\max} = (p_{\max}-e_{\max})^2,\]
where $p_{\max}$ is the probability assigned to the maximum probability category, and $e_{\max} = 1$ if the observation falls into that category and $0$ else. 
Unfortunately, it turns out that this score is *improper*: it does not reward calibration and accuracy. Let us look at an example forecast for just one gridpoint:

```{r,echo = FALSE}
knitr::include_graphics('example_plot.png')
```

In this example, we compare a near-climatological forecast (red) with a prediction issued by a forecaster (blue). The highest probability categories are indicated by the shaded area: for the forecaster it is the 'above normal' category, for the climatology-forecast the 'below normal' category. Below the figure, the scores achieved by the forecaster and climatology are shown for all three possible outcomes.
The climatology gets a better (lower) Brier score when the observation is 'normal' or 'above normal'. This is paradoxical, since the forecaster assigned higher probabilities to these categorie. This highlights the improperness of the max-Brier score: When evaluating predictions with this score, the best forecast does usually not get preferred.

This is unintuitive, because the (standard) Brier score is proper. However, the Brier score is designed for predictions of two-category-events *with fixed categories*. In the definition of $BS_{\max}$ the categories are 'highest probability category' vs. the rest. Therefore, the two categories *depend on the forecast probabilities* and therefore may vary between different predictions. This makes the Brier score improper.

However, a nice application of Theorem 1 of [this](https://arxiv.org/abs/1506.07212) paper shows that there is a class of proper scoring rules that can be evaluated, when only the probability of the most likely category is known. For example, we can use the score
\[ cBS_\max:= p^2_{\max} - 2p_\max e_\max + 1.\]
Note that this score satisfies
$cBS_\max=BS_{\max} - e_\max +1$, so it's a corrected version of the max-Brier score which is proper and avoids the problems above. Adding $+1$ in the definition of the score is not necessary but convenient: it ensures that the score is nonnegative and a perfect score is 0.

Usually we want to know whether our prediction outperformed climatology. For most scores we can consider skill scores, but unfortunately this does not work here. Climatology assigns to all three categories *equal* probabilities (1/3), and therefore does not really have a maximum-probability-category. Thus, the definition of $e_\max$ makes no sense for a climatological forecaster. However, a reasonable viewpoint is that for a climatological forecast the maximum-probability-category can be picked at random, since all categories are getting assigned the same probability.
This means that climatology achieves a score of 4/9 with probability 1/3 (when $e_\max = 1$), but only achieves a score of 10/9 with probability 2/3. Thus, on average the climatological forecast achieves a score of $\frac 1 3 \frac 4 9 + \frac 23 \frac {10}9 = \frac{24}{27}$. A forecast that attains a $cBS_\max$ of below 24/27 performs on average better than climatology. We therefore define the 'skill score'
\[cBSS_\max := 1 - \frac{27}{24}cBS_\max.\]
Note that this is not a skill score in the strict sense, but can be interpretet similarly: values above 0 indicate higher skill than climatology on average, with a $cBSS_\max$ of 1 corresponding to a perfect forecast.

To try this out in action, let us look at the 2021 tercile forecasts

```{r,fig.height = 4,fig.width = 12}
# data_dir = '/nr/project/stat/CONFER/Data/validation/example_data/202102/' # as in section 3
fn = 'Ens_Prec_1monLead_MAM_Prob_EnsRegrCPT-avg.nc'

dt = netcdf_to_dt(paste0(data_dir,fn))
dt = dt[!is.na(below) | !is.na(normal) | !is.na (above)]

p1 = ggplot_dt(dt,data_col = 'below', midpoint = dt[,min(below,na.rm = TRUE)])
p2 = ggplot_dt(dt,data_col = 'normal', midpoint = dt[,min(normal,na.rm = TRUE)], high = 'darkgoldenrod') # see https://www.r-graph-gallery.com/ggplot2-color.html for an overview of color names.
p3 = ggplot_dt(dt,data_col = 'above', midpoint = dt[,min(above,na.rm = TRUE)], high = 'darkgreen')

ggarrange(p1,p2,p3,ncol = 3)
```

*In order to evaluate these forecast the high resolution CHIRPS-data of the past is missing!*

```{r}
fn = "PredictedProbabilityRain_Mar-May_Feb2021_new.nc"
dt = netcdf_to_dt(paste0(data_dir,fn))

dt[,normal := normal/100][,above := above/100][,below := below/100]
```



## Evaluating cross-validation predictions {#cv-eval}

Here we evaluate the cross-validation data prepared in Section \@ref(cv-data).
The data table contains observations for past years for the FMA season, along with 'best-guess'-predictions, meaning that they are single numbers, not probabilities:
```{r}
print(dt_cv)
```
Such predictions are often called *point forecasts*, whereas forecasts specifying probabilities are called *probabilistic*. 
We already have the data in the shape we want it to be, containing both predictions and observations as one column each. Let's have a look at the bias in our predictions:

```{r}

### check out local biases:
bias_dt = dt_cv[,.(bias = mean(prediction - observation)), by = .(lon,lat)] # grouping by lon,lat, and season means that the mean is taken over all years.
bias_dt[,range(bias)] # get an idea of the range for plotting

rr = c(-15,15) # fix range, to make plots comparable

 ggplot_dt(bias_dt,
           data_col = 'bias', 
           rr = rr, # fix range to make it comparable to pp2
           mn = 'bias of FMA prediction',
           midpoint = 0)

```

For evaluating the locally varying skill of these predictions we can calculate the average square error per location and compare it to the square error we would get by predicting local climatology, i.e., just the mean precipitation for that location. This is called the mean square error skill score (MSES) and is calculated by the function `MSES`.
For fair comparison, the climatology needs to be calculated leave-one-year-out: For example, a climatological prediction for 2021 prediction would be the mean precipitation over all years *except 2021*. Otherwise, the climatological forecast uses information that would not be available in a prediction setting, which inflates the skill of the climatology forecast. 


```{r,  warning=FALSE}
### analyze mean square error skill scores
msess = MSES(dt_cv,
             f = 'prediction', # column name of forecasts
             o = 'observation', # column name of observations
             by = c('lon','lat')) # the skill scores should be computed for each location separately

# plot results:
ggplot_dt(msess,
          data_col = 'MSES', 
          midpoint = 0, # center color scale, white is 0
          mn = 'MSE skill score, FMA')

```
As for all skill scores, positive values indicate that the prediction has higher skill than climatology, negative values indicates lower skill. Skill scores are moreover 'standardized' such that a score of 1 corresponds to a perfect forecast. 
Most evaluation functions in `SeaVal` require similar input as `MSES`:
* `dt`: a data table containing both observations and predictions.
* `f`,`o`: the column names of the predictions and observation.
* `by`: the column names for dimension variables to group by.
[Add more context here]
Note that there is also a (faster) function `MSE` if we're not interested in skill scores, but simply want to compute MSEs. Both function can also handle ensemble predictions, see function documentation.
If we want to analyze results by countries, we can use the function `add_country_names` that adds a column with country names to the data table:

```{r}
# check out average MSEs and MSESSs per country:
msess = add_country_names(msess)
print(msess)
msess_by_country = msess[,.(MSE = mean(MSE),
                            MSES = mean(MSES)), by = country] # take averages by country

print(msess_by_country)

```


Skill scores strongly depend on the skill of the climatological prediction, see Section \@ref(eval-ex-pr). This makes it somewhat problematic to average them in space, as skill scores for different grid points with different climatologies have different meanings. A more appropriate way to see whether the prediction outperformed climatology on average for a given country is by considering average score differences:

```{r}
# positive values indicate better performance than climatology:
msess[,.(score_diff = mean(clim_MSE - MSE)),by = country]
```

The MSE (and its associated skill score) penalizes both systematic forecast errors (i.e. biases) and non-systematic forecast errors. The latter are a consequence of general forecast uncertainty and there is no easy way to reduce them. Biases, however, can often be removed through statistical post-processing, and it is therefore interesting to consider measures for forecast performance that penalize only non-systematic forecast errors, thus giving an idea of the *potential skill* of a forecast system.

The standard metric to assess the *potential skill* is the *Pearson correlation coefficient (PCC)*. This is the usual correlation coefficient where forecasts and observations are standardized by their respective climatological means and standard deviations, and then the average product of these standardized variables is calculated. The function `PCC` performs these calculations and is used in the same way as `MSES` above.

```{r,warning = FALSE}
### calculate Pearson correlation coefficients
pcc = PCC(dt_cv,
          f = 'prediction', # column name of forecasts
          o = 'observation', # column name of observations
          by = c('lon','lat')) # the correlation coefficient should be computed for each location separately

# the maximal range for a correlation coefficient is [-1,1], but sometimes it is useful to narrow it:
rr = c(-0.75,0.75)

ggplot_dt(pcc,
          data_col = 'rho', 
          rr=rr,
          mn = 'Pearson correlation coefficient, FMA')
```


While there is no technical requirement that the forecasts and observations follow a particular probability distribution when the Pearson correlation coefficient is employed, this metric is best suited for continuous distributions (i.e. it is unlikely to encounter duplicate values) that are relatively symmetric around the mean. For shorter (e.g. weekly) accumulation periods and in dry climates, the distribution of precipitation usually becomes rather skewed and contains a number of zeros. A new metric, the coefficient of predictive ability (CPA), has recently been developed and constitutes an excellent alternative to the PCC as a measure of potential forecast skill in that situation of strongly asymmetric distributions with multiple identical values. See [here](MotivationCPA.pdf) for more background information about the CPA. The function `CPA` performs the calculations and is used in the same way as `MSES` and `PCC` above.

```{r,warning = FALSE}
### calculate coefficient of predictive ability
cpa = CPA(dt_cv,
          f = 'prediction', # column name of forecasts
          o = 'observation', # column name of observations
          by = c('lon','lat')) # the CPA should be computed for each location separately

# the maximal range for the CPA is [0,1]
# a value of 0.5 corresponds to no skill (more details can be found in the document under the link given above) 
rr = c(0,1)

ggplot_dt(cpa, 
          data_col = 'cpa', 
          rr=rr,
          mn = 'Coefficient of predictive ability, FMA')

```

Just like the MSES, PCC and CPA can be averaged by country using the function `add_country_names`:

```{r}
# check out average PCCs and CPAs per country:
pcc = add_country_names(pcc)
cpa = add_country_names(cpa)

pcc_by_country = pcc[,.(rho = mean(rho)), by = country]
cpa_by_country = cpa[,.(cpa = mean(cpa)), by = country]

print(pcc_by_country)

print(cpa_by_country)
```



## Exceedence probabilities {#eval-ex-pr}

Another forecast product issued at GHACOFs are exceedence probabilities of precipitation for certain thresholds, generally related to crops important for the region.
A proper scoring rule based on the predicted exceedence probability $p_\text{exc}(c)$ of a threshold $c$ is the Brier score of exceedence
\[BS_{ex}(c) := (p_\text{exc}(c) - 1\{y>c\})^2,\]
where $1\{y>c\}$ equals 1 if the observation $y$ exceeded threshold $c$, and 0 else. Skill scores for comparison with a climatological forecast can be calculated in the usual way. The climatological forecast for the exceedence probability is the fraction of past observations that exceeded the threshold. In Section \@ref(data-ex-prexc) we already derived this dataset:

```{r}
print(dt_prexc)
```
This dataset contains predictions of exceedence (by different models) for several thresholds (rthr), as well as observed rainfall and a climatological prediction for the exceedence probabilities. This is everything we need to compute the $BS_{ex}$-skill score. To this end, we have the function `BSS_ex_dt`. If we would not have a climatological exceedence forecast available, we could have still computed the $BS_{ex}$-score using the function `BS_ex_dt`. This is still usefull for comparing competing models (see below), but does not tell us where the prediction is better or worse than climatology. 

```{r,plot.width = 10,plot.height = 12,cache=TRUE}
# bss_dt = BSS_ex_dt(dt_prexc,
#                    f = 'pexcd',
#                    threshold_col = 'rthr',
#                    o = 'prec',
#                    by = c('model','month','lon','lat'))
# 
# print(bss_dt[!is.na(BS_ex)])
```

Skill scores are generally not defined when the climatological prediction is perfect and the climatological score is zero. This happens frequently for exceedence probabilities (e.g. lines 3 and 4 in the data table above) at locations where the considered threshold has never been exceeded in the observation. Simply for plotting reasons we put the skill score to -1 in this case if the prediction scores above 0 (since the climatological prediction wwas better in this case), and to 0 if both climatology and prediction assign a probability of 0. 
Let us look at the skill scores by the different models for the March forecast for exceedence level 200mm:
```{r}
# make a list of skill score plots:
# theme_set(theme_bw(base_size = 10)) # smaller font
# plot_list = list()
# for(mod in unique(bss_dt[,model])) # 1 plot for each model
# {
#   plot_list = c(plot_list,list(ggplot_dt( bss_dt[model == mod & month == 3 & rthr == 200],
#                                           'BSS_ex',
#                                           mn = mod, 
#                                           high = 'red',
#                                           midpoint = 0,
#                                           rr= c(-1,1),
#                                           guide = guide_colorbar(title = NULL, barwidth = 75, direction = 'horizontal'))))
#   
# }
# 
# ggpubr::ggarrange(plotlist = plot_list,ncol = 3,nrow = 3,common.legend = TRUE,legend = 'bottom')

```
Here, red color indicates better performance of the prediction than climatology. Large areas of the map are blue, which indicates better performance of the climatological forecast than of the prediction models. However, these are mostly areas where the observations never exceeded 200mm. Therefore, the climatological forecast issued a 0% chance of rainfall exceeding 200mm, whereas all actual prediction models issued a small positive probability and therefore performed 'worse'. This not so much highlights a problem of the forecasts than rather a problem of skill scores, which become degenerate whenever the climatological prediction is near perfect.

For comparing overall performance, we can average scores spatially. Note that, because of the above-mentioned effect, it is important not to average skill scores. However, since we have a climatology forecast in our data table, we can compute a spatially averaged score for climatology as well. Thus, we can compare whether the prediction models performed on average better or worse than climatology.

```{r}
# mean_scores = bss_dt[,.(BS_ex = mean(BS_ex,na.rm = T)),by = .(model,month,rthr)]
# # get climatology score as well:
# mean_clim_score = bss_dt[model == model[1],.(BS_ex = mean(clim_BS_ex,na.rm = T)),by = .(month,rthr)]
# mean_clim_score[,model := 'clim']
# 
# mean_scores = rbindlist(list(mean_scores,mean_clim_score),use.names = TRUE)
# print(mean_scores)

```

Here, every model gets assigned a single mean score for each month and each threshold (the mean score over all gridpoints). Lower values indicate better overall performance. Let us plot the data:

```{r, plot.width = 10}
#pp = ggplot(mean_scores) + geom_line(aes(x = month,y = BS_ex,color = model,linetype = model)) + facet_wrap(~rthr,nrow = 1)
#
#print(pp)
```

The plot shows that, averaging over all grid points, a climatological forecast does much better than all the systems, which probably indicates that the systems need to be bias corrected.


## Temperature

In our folder of example data we also have a file containing temperature predictions.
The file already contains correlations as well. Here we simply visualize these correlations as plots:

```{r,fig.width = 10,fig.height = 7}

fn = 'TrefEnsRegr_monthly.nc'
dt = netcdf_to_dt(paste0(data_dir,fn))

# plot correlations of predictions for all five models at all lead_times:
# create list of plots:
plot_list = list()
for(leadtime in 1:3)
{
  for(mod in 1:5)
  {
      plot_list = c(plot_list,list(ggplot_dt(dt[model == mod & lead == leadtime],
                                        'corr',
                                        rr = c(-1,1),
                                        mn = paste0('model ',mod,', lead ',leadtime),
                                        discrete_cs = TRUE,
                                        binwidth = 0.2,
                                        guide = guide_colorbar(title = NULL, 
                                                               barwidth = 75,
                                                               direction = 'horizontal'))))   }  
}


#plot as grid:
do.call('ggarrange', c(plot_list,ncol = 5,nrow = 3,common.legend = TRUE,legend = 'bottom'))

```
